requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
blinker==1.7.0
cachelib==0.13.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
Faker==24.14.0
Flask==3.0.0
Flask-Cache==0.13.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
msgspec==0.18.6
pearhash==1.0.1
pillow==10.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.2
